feat: made separate folder for all utils
